Remove usages of exotic primitives where applicable Use 'int' primitive in place of 'short' and 'byte' wherever possible. Fixes out-of-range issues for values greater than short.MaxValue.

+++ EOBot/BotHelper.cs
@@ -30,7 +30,7 @@ namespace EOBot
             var accParams = new CreateAccountParameters(name, password, password, name, name, name + "@eobot.net");
             var nameResult = await accountActions.CheckAccountNameWithServer(name);
             return nameResult >= AccountReply.OK_CodeRange
-                ? await accountActions.CreateAccount(accParams, (short)nameResult)
+                ? await accountActions.CreateAccount(accParams, (int)nameResult)
                 : nameResult;
         }
